Hud is fix
In DrawManager:
Removed the Delete since m_window is declared on the stack in Engine,
NOT DYNAMICALLY. Also because it caused crash on shutdown.

In gamestate Constructor:
Declaring text and circles and setting the score to 0

In gamestate Update:
Check if monster died, increase m_score by 100. Then convert and set the
score int to string so we can display it with m_scoreDisplay.

In gamestate Draw:
Draw the circles(lives) in a for loop. Change the second for loop
parameter from 3 to number of current lives when lives are implemented.
Circle should be changed to better icons later.
Draw the scoreDisplay

In GameState(.h):
Added a font, circle, text and int to use for HUD

In InputManager:
Added a constructor and destructor since the program crashed on
shutdown.